CodeOwnerResolver: Allow to set the user for which the visibility is checked
CodeOwnerResolver filters out code owners that are not visible to the
current user, however when validating code owner configs on submit we
want to do the visibility check for the uploader of the patch set which
can differ from the current user (user who is doing the submit). To
support this CodeOwnerResolver now allows to specify an account for
which the visibility is checked (whether this account can see the
accounts of the code owners).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ic89230e44179b33f528ce4571805211a6ed5ed81