Validate code owner config files on submit
We already validate code owner config files on push, but this is not
sufficient since open changes with invalid code owner config files may
have been created before the code-owners plugin was installed or before
the code owners functionality was activated for a project or branch. In
addition it is possible that configuration parameters have been changed
after a change with code owner config files was uploaded and before the
submission is attempted (e.g. the configuration parameter that controls
the account visibility has an effect on the validation).

The validation that is done on submit is exactly the same validation
that is also done on upload. This means account visibility checks are
done from the perspective of the uploader (not the submitter).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I4f8726667b0e73415f9d0f572fdf3966dc057743