Improve messages for code owner config validation
When validating code owner configs for commits we can have the following
cases:

1. the code owner config file validation doesn't need to run because
   the commit doesn't touch code owner config files
2. the code owner config file validation run and no issues were found
3. the code owner config file validation run and only warnings were
   found
4. the code owner config file validation run and errors were found
5. the code owner config file validation was skipped because the code
   owner config functionality is disabled
6. the code owner config file validation couldn't run (due to an invalid
   plugin configuration)

At the moment these cases result in the following messages:

1. no message
2. no message
3. warning messages
4. summary error message ("invalid code owner config files") + error
   messages
5. no message
6. no message

So far only the error case had a summary message (the message of the
exception that is thrown if there are errors). Now we allow to provide a
summary message also in the other cases. Since no exception is thrown in
these cases we sent it to the client as another validation message which
is sent first. The type of the message is HINT, unless there are
warnings in which case it is WARNING.

With this in place we improve the messages for the cases listed above:

1. [UNCHANGED] no message
2. summary hint message saying that the code owner configs were
   validated, but no issues were found
3. summary warning message ("invalid code owner config files") + warning
   messages
4. [UNCHANGED] summary error message ("invalid code owner config files")
   + error messages
5. summary hint message saying that the validation is skipped + hint
   message saying that the code owners functionality is disabled
6. summary warning message saying that the validation is skipped +
   warning message saying that the code owners configuration is invalid

In a follow-up change we will add another case:

7. the commit is rejected because it touches code owner config files but
   code owner config files are configured to be read-only

Since we can now control the summary message (instead of hard-coding
"invalid code owner config files"), we can return the following messages
in this case:

7. summary error message saying that code owner config files cannot be
   edited due to the read-only configuration + error message per touched
   code owner config file

Change-Id: I47188c8ebd7ec6f9d7f53ed2ba7b0bae343e0ae6