Document how to setup the plugin
The code-owners plugin provides quite some configuration parameters to
adapt the functionality to the needs of individual projects and teams.

For some configuration parameters it is important to set them before
enabling the code owner functionality.

To make host and project owners aware of the configuration parameters
that are important, this change adds a setup guide to the documentation.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ifcd046f03cd790d20f5121dd452200e06b054379