Document responsibilities of author & code owners on renames
When files are renamed it is possible that the ownership of the current
code owners is lost, e.g.:

  per-file *.md=foo@example.com

If now 'config.md' gets renamed to 'config.txt' user foo is no longer a
code owner of this file. In this case it's the responsibility of the
author doing the rename and the current code owners to ensure that the
file still has proper code owners after the rename.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ia3e9e798de54b958e7a771572b26de86c8934ce4