Filter out code owners that cannot see the branch
Code owners that cannot see the branch cannot approve paths in this
branch. Hence returning them to the client is not useful.

For changes the visibility check is done for the destination branch,
rather than the change. This allows to add code owners as a reviewer to
private changes, that couldn't see the change before they were added as
a reviewer (also see comment in AbstractGetCodeOwnersForPath).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I80c44d96156e239750c2e7de54b1ca2b9cd394d2