Add REST endpoint to list the code owner config files in a branch
This REST endpoint allows external services to find all code owner
config files in a branch, e.g. to perform updates on them.

This REST endpoint scans all code owner config files in the branch and
it is expected that it can take a long time if the branch contains many
files. This is why this REST endpoint must not be used in any critical
paths where performance matters.

In the BranchCodeOwners API class we have an inner class to represent
the request to list code owner config files. With the current
functionality having this class is not needed and we could just let the
codeOwnerConfigFiles() method return the paths directly, but it allows
us to add parameters to this REST endpint later without breaking the API
and without needing to introduce additional methods in BranchCodeOwners.
E.g. in a follow-up change we want to add an email parameter that limits
the returned code owner config files to those that contain the given
email.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If8993f215c6e005aa3c11adfbe21b0b5dd28b5f7