Add REST endpoint to list the code owner config files in a branch
This REST endpoint allows external services to find all code owner
config files in a branch, e.g. to perform updates on them.

This REST endpoint scans all code owner config files in the branch and
it is expected that it can take a long time if the branch contains many
files. This is why this REST endpoint must not be used in any critical
paths where performance matters.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If8993f215c6e005aa3c11adfbe21b0b5dd28b5f7