CodeOwnerProjectConfigJson: Inject CodeOwnersPluginConfiguration
This way callers do not need to get CodeOwnersPluginConfiguration
injected and then pass all config parameters into
CodeOwnerProjectConfigJson.

This requires using mocking for the CodeOwnerProjectConfigJsonTest but
that's fine. In return we get rid of the tests that check for null
inputs. They are no longer needed since callers can no longer pass in
null values.

Change-Id: I75d6d393e788ec5f5224576b1fa9f8fd1f0fb973
Signed-off-by: Edwin Kempin <ekempin@google.com>