Make it configurable whether implicit code owner approvals are assumed
So far a code owner approval from the last uploader has been assumed if
the uploader was a code owner. Now this behaviour is configurable and
needs to be switched on explicitly by setting
plugin.code-owners.enableImplicitApprovals in gerrit.config or
codeOwners.enableImplicitApprovals in code-owners.config on
project-level.

By default we disable implicit approvals as this is the safer option. As
documented by change I07a9a9c21 implicit approvals can be an issue if
the code owners are not aware of it, e.g. if a contributor pushes a
change to a wrong branch and a code owner helps them to get it rebased
onto the correct branch, the rebased change has implicit approvals from
the code owner, since the code owner is the uploader, which may not be
wanted.

If implicit code owner approvals are disabled, code owners can still
self-approve their own changes by voting on the change.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Icf52ffb44d9d85b5b4d688a3ed5643b659345280