Allow to define global code owners
Some projects like Android consist out of many repositories and many
branches. If a bot needs to be a code owner of all files (in all
repositories and in all branches), the bot would need to be added as a
code owner in the root OWNERS files of all the repositories and
branches. This is not only a big amount of work to setup, but is also
fragile. E.g. it's easy to miss assigning code ownership in some
repositories/branches, local project owners may remove the ownership
accidentally, you need to remember doing this configuration whenever a
new repository or branch gets created. For some projects (Android) a
misconfiguration can result in a severe outage. Hence to make the
configuration of global code owners easier and more reliable we allow to
define users that should own all files across repositories and branches
in gerrit.config. As with all configuration settings, it's possible to
override this setting on project-level, in case some projects should be
exempted or have different global code owners.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ic08fe297b8d01b708d192e7baf2534ebb93805f5