CodeOwnerConfigScanner: Allow to handle invalid code owner config files
At the moment the CodeOwnerConfigScanner would just fail if it hits any
invalid (aka non-parseable) code owner config file, which is bad.

We could simply ignore them, but it's better to add a callback for
invalid code owner config files so that callers can handle them if they
want.

For now all callers ignore invalid code owner config files. However in a
follow-up change we will implement a REST endpoint to validate the code
owner config files in the repository and thanks to the callback this
REST endpoint will now be able to also report invalid code owner config
files.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I9058f4f5aace7e9fe739fb03f47a33d4099b1eed