Support to validate the code owner config files in a revision/patch set
This allows to validate the code owner config files in open changes. The
validation is performed from the perspective of the uploader, so that
the validation is exactly the same as the validation that will be done
on submit.

At the moment this REST endpoint doesn't provide much value since code
owner config files are already validated on push and the upload is
rejected if the commit contains invalid code owner config files. So in
practice changes should not contain invalid code owner config files.
However in future we want to allow project owners to switch off the
validation on upload and instead run a bot that does the validation on
the open changes as a pre-submit check. In this case the bot would use
this REST endpoint to do the validation on the open changes.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I010fbc9804bf14a418572234626c5a69e75840f1