CodeOwnerConfigFile: Pass in RevWalk from which the revision should be loaded
If a revision is newly inserted, but not committed to the repository
yet, it can only be seen from the RevWalk instance that created it, but
not by newly instantiated RevWalks. So far we used
VersionedMetaData#load(Project.NameKey, Repository, ObjectId) to load
CodeOwnerConfigFile when a revision was given, but this method creates a
new RevWalk instance to lookup the revision and hence it would fail with
MissingObjectException if the revision was newly inserted and not
committed yet. To support loading CodeOwnerConfigFile from newly
inserted revisions we now allow to pass in the RevWalk instance from
which the revision should be loaded. This means if CodeOwnerConfigFile
should be loaded from a certain revision we require a RevWalk and the
revision as input now. If the CodeOwnerConfigFile should be loaded from
the current revision we only need a Repository instance. Since both
cases require different parameters now, CodeOwnerConfigFile has separate
methods for them now, load and loadCurrent.

This is only a first step to support loading CodeOwnerConfigFile from
newly inserted revisions, that have not been committed yet. Follow-up
changes will change the higher layers so that we can pass through the
RevWalk instance that should be used to load the revision.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I1943db7cf1adca0019d229ecc2959bf0161235ff