CodeOwnerBackend: Pass in RevWalk from which the revision should be loaded
If a revision is newly inserted, but not committed to the repository
yet, it can only be seen from the RevWalk instance that created it, but
not by newly instantiated RevWalks. To support loading code owner
configs from newly inserted revisions that have not been committed yet,
we now allow to pass in a RevWalk instance that should be used to load
it.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ie34eb26ce7fcfb57e3dcf6a7d649c376476b8330