ChangedFiles: Pass in RevWalk from which the revision should be loaded
If a revision is newly inserted, but not committed to the repository
yet, it can only be seen from the RevWalk instance that created it, but
not by newly instantiated RevWalks. To support loading changed files
from newly inserted revisions that have not been committed yet, we now
allow to pass in a RevWalk instance that should be used to load it.

In addition we allow to pass in the repository config, since callers
that will need to pass in the RevWalk (CodeOwnerConfigValidator) will
have the repository config already available, and hence we will not need
to reopen the repository just to read the repository config.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: If80c7fcd48219b7217de5d517fcd817cf974e999