Allow to disable the validation of code owner configs on commit received
Instead of rejecting commits with invalid code owner config files
directly on upload, it may be preferred to allow the upload and then do
the validation on the open changes as part of a pre-submit validation.
To make this possible we support to disable the validation of code owner
config files when a commit is received.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I31a38af249c9d80491a3ce7da8157f7b606cabfd