CodeOwnerApprovalCheckTest: Make the approvedByGlobalCodeOwner more readable
Looking at the test one could think that the file was already implictly
approved by the admin user who creates the change and is a code owner.
This is not the case since implicit approvals are disabled by default.
However the test becomes more readable if we use a non-code-owner user
to create the change. Add a verification that the file is not approved
before the global code owner adds their approval. Also add some more
comments to make the code easier to read.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ib626315a1ca5efdd9b2e5a066f878507606d1ad8