CodeOwnerApprovalCheck: Return 409 Conflict if destination branch was deleted
It can happen that the destination branch of a change gets deleted. In
this case we cannot compute the code owner status for the change, since
the code owner configs are loaded from the destination branch, which is
not possible when the branch was deleted. Return 409 Conflict in this
case rather than failing with an internal server error.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ibd694fc22113dad89c50cf8404f25b703b79855d