Return 409 Conflict if a code owner config is invalid
At the moment we fail with 500 ISE if there is a code owner config, but
it's not an error that should be fixed by the Gerrit team. Rather it's a
configuration issue that needs to be fixed by the project owner.
Returning 409 Conflict is consistent with have Gerrit core behaves in
similar situations (e.g. when there is an invalid Prolog rule).

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I27bdcf994cc9a0e086f02706d729594ac7d1a3a2