Allow to configure a URL that explains how to get a code owner override
Sometimes a change needs to be submitted urgently and there is no time
to gather the required code owner approvals. In this case a code owner
override can be applied to make the change submittable. Usually the
permission to apply an override is limited to a small set of super
users that may not be known to the user needing a code owner override.
To allow users to find a person who can grant a code owner override we
allow to configure a URL to a documentation page where it's explained
how an override can be retrieved. This documentation page is sepcific
for a host or project. The frontend gets the URL via the
GetCodeOwnerProjectConfig REST endpoint and can display a link or icon
for it.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ice075c75f379491ded97d1f0dd8d06a942cd4188