GetCodeOwnerProjectConfig: If disabled set only status.disabled
If the code owners functionality is disabled for the project
(status.disabled = true) then the other config parameters are not
relevant and do not need to be returned.

Trying to return them anyway is bad, because they may be invalid and
then the request cannot succeed. E.g. at the moment
GetCodeOwnerProjectConfig fails with 500 Internal Server Error if a
non-existing label is configured as required approval. The required
approval configuration is not relevant while the code owners
functionality is disabled and should not cause the request to fail.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ifd2cce2ce973a35b12dc855b8c074d0a3d981a62