GetCodeOwnerProjectConfig should not fail with 500 ISE if config is invalid
If the plugin configuration is invalid we want to fail with 409 Conflict
rather than 500 Internal Server Error. This is ensured by
InvalidPluginConfigurationException.ExceptionHook which returns
409 Conflict if an exception is caused by
InvalidPluginConfigurationException, but there was one case where an
invalid configuration didn't throw this exception and hence the request
failed with 500 ISE. This was the case if the default required approval
(Code-Review+1) was used but the project didn't have the Code-Review
label.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I76e6de49b089f211a03a81c52430296806b8fe5d