Allow running the code owner config validation as dry-run
When rolling out the code-owners plugin, twice there were issues with
the code owner config validation, which is why it got disabled now. Add
a dry-run option so that we can test it before, turning in on for real
again.

This change is compatible with existing configurations which have
already set a value for the enableValidationOnCommitReceived config
parameter.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I22decdec71936f416fae1da5cd1530a3b2992e1b