Add class that updates all code owner configs in a branch atomically
CodeOwnerConfigFileUpdateScanner scans a branch for code owner config
files and allows a callback to update them. The callback gets the plain
file content, which means that the code owner configs are not parsed.
Hence the callback also gets invoked for non-parseable code owner
configs. Updates to the code owner configs are committed atomically with
a single commit. In case of a no-op the branch is not updated.

In a follow-up change we will make use of
CodeOwnerConfigFileUpdateScanner to support replacing an email in all
code owner config files of a branch.

Change-Id: I2113ca172060ddfc97d548ddbbecef19aca5a7e2
Signed-off-by: Edwin Kempin <ekempin@google.com>