GetCodeOwnerBranchConfig: Include a flag if no code owner config file exists yet
If a branch doesn't contain any code owner config file yet, we are in
bootstrapping mode where project owners count as code owners. Let the
frontend know about this, so that it can inform users if this is the
case. E.g. this may help users to understand why no code owners can be
suggested.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I84560336f862c0d2d8650b914732ff60016b2f72