Document that imports and global owners are always read from the current revisions
The GetCodeOwnersForPathInBranch REST endpoint allows to specify a
revision from which the code owner config files in this branch should be
read. Clarify in the documentation that imports from other branches or
repositories as well as global code owners from refs/meta/config are
still read from the current revisions.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I968977a0d1214f59804234367f0a31f202a3df84