Use new Get Code Owner Branch Config REST endpoint in frontend
At the moment the code-owners frontend is using the Get Code Owner Project Config REST endpoint to read the plugin configuration and to decide whether the code owner UI controls should be shown.
It turned out the this REST endpoint is too slow if a repository contains many branches, as it returns some information (whether the code owners functionality is enabled/disabled and which code owner backends is used) for every branch. Computing this information for all branches is unnecessary as the frontend is only interested in the information about the target branch of the change.

Change-Id: I357f0664fcfb0a06a690312766e7fe057873a348