Count sticky approvals on old patch sets as code owner approvals
If an approval was applied on an old patch set which is still valid for
the current patch set because it is sticky we must consider it as code
owner approval the same way as approvals that have been applied on the
current patch set.

The computation of sticky approvals is done by ApprovalInference which
we are now using by retrieving the approvals through ApprovalUtil.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I2131162f5de234a53ba923c9c8111a13d7a555ff