Extract ExceptionHook implementation into an own class
Originally the ExceptionHook implementation only cared about the
exception that is caused by an invalid plugin configuration, hence it
was fine having it as inner class of
InvalidPluginConfigurationException. But by now it's also handling the
exception that is caused by invalid code owner config files and those
are not part of the plugin configuration. Hence it's better to have a
separate class for this now.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: I8a0ebfbd3eea80981993cc8bf283e575709a1203