Introduce CodeOwnersModel and ModelLoader and use them in UI elements
This change removes direct dependencies between CodeOwnersService
and UI elements. Instead, UI elements displays/uses model as a
data source and calls ModelLoader when a model property should be
populated with a data.
This change doesn't update suggest-owners.js (it displays the
list of suggested owners). The suggest-owners.js is updated in the
upcoming change.

Change-Id: I9d4c4fda6eac4aab674a9928db9f1c408f93b8f3