Handle restAPI errors
The pluginStatus property is added to the CodeOwnersModel class. The
ModelLoader catches all excpetions and setup failed pluginStatue when
a request fails.
If a fail happens, the error is displayed:
* At the top of page
* Near the "Code owners" label

The banner is required due to the following reason: If the code-owners
plugin configuration has a problem, the "Code owners" label can be
hidden (backend doesn't return the label). But we still want to display
a error message, so the code uses banner to show the error.

Change-Id: I0701b269ae32d2727e2749324a664f071eec9c80