Don't filter out service users when listing code owners for path in branch
Currently we have 2 REST endpoints to list code owners for a path that
behave the same:

* GetCodeOwnersForPathInBranch
* GetCodeOwnersForPathInChange

Going on GetCodeOwnersForPathInChange will be used to suggest code
owners for a path on a change, whereas GetCodeOwnersForPathInBranch
lists all code owners for path. The difference is that for
GetCodeOwnersForPathInChange we will filter out code owners that should
not show up in the suggestion (e.g. service users) whereas
GetCodeOwnersForPathInBranch includes all code owners.

The frontend part was already adapted to use
GetCodeOwnersForPathInChange for suggestion code owners (see change
Ie04e60913).

GetCodeOwnersForPathInBranch is not used by the frontend, but is
intended to be used by bots that want to query all code owners.

Signed-off-by: Edwin Kempin <ekempin@google.com>
Change-Id: Ifbec40349fa1bde36ea36822e6f4bc2a39fd295e